Fix #3: use one socket connect per request
Previously all requests shared the same connection which meant that the client
had to wait that a request finished until it could issue another request.
Otherwise, replies could end up for the wrong request causing all kinds of
problems. This synchronous request-reply model could not be kept because of
inherent asynchronicity of certain mechanisms (e.g. grab + software trigger).

This change removes the always-open socket connection and replaces it with a new
connection per request. Due to the OS multiplexing, replies cannot be mistaken.