ucad will error on Python-supported libuca
becuase the run_callback runs in a separate thread and we'd need to use
Python subinterpreter which has caveats. At the same time, Python with
ucad is just not needed since ucad is a standalone C program which uses
the cameras without any Python, so GIL manipulation is not needed in the
first place. This will require libuca >= 2.4.1