Adds the properties 'mirror' and 'rotate' to all remote cameras
The properties are send with the image metadata, telling the receiver if the frames should be transformed.
This would be a python implementation for the 'standard' transformation:
def _convert_image(image, mirror: bool, rotate: int):
   if mirror:
      image = np.fliplr(image)
   if rotate:
      image = np.rot90(image, k=rotate)
   return image

Adds the properties mirror and rotate to all remote cameras

The properties are send with the image metadata, telling the receiver if the frames should be transformed.
This would be a python implementation for the 'standard' transformation:
def _convert_image(image, mirror: bool, rotate: int):
   if mirror:
      image = np.fliplr(image)
   if rotate:
      image = np.rot90(image, k=rotate)
   return image